basic implementation of stealing a card when played three kittens, cards with stealing function are now extending a superclass
